feat: Modify routes
- Add home route
- Refactor history page
