+++ src/pages/HomePage.tsx
@@ -1,0 +1,3 @@
+export default function HomePage() {
+  return <div>Home</div>;
+}
